Propose a fix for globalsign/mgo#120
As discussed in the issue globalsign/mgo#120, isMaster() can cause a
deadlock with the topology scanner if the connection it makes dies
before running the command; mgo automagically attempts to make another
socket in acquireSocket, but this can't work without topology.

This commit forces isMaster() to actually run on the intended socket.